Refactor header component for improved navigation
Restructured `header.component.html` to enhance the desktop and mobile menus with conditional rendering based on `showToggle`. Updated icons for the search button and added buttons for dark mode and mobile filters. Improved the notification dropdown to display a dynamic list of notifications, and refined the language and profile dropdowns.

In `header.component.ts`, added `Router` to imports and replaced the `link` property in the `profiledd` interface with an `action` method for dynamic navigation. Introduced an `authService` for user authentication and added a `signOut` method to handle user logout.